requirements.txt
##### Deep learning packages
#pytorch 
torch
torchvision
tensorflow
torchviz
albumentations
ensemble-boxes
tensorboard
scikit-learn 

############ Firebase packages
pyrebase
requests>=2.26
scp
pynput

######## Web API packages
fastapi 
uvicorn
python-multipart

##### GUI packages
PyQt5

##### preprocessing packages
Pillow
# opencv-python
opencv-python-headless 
opencv-contrib-python==4.1.2.30
scikit-image>=0.19.1
imutils
# pylops
polyline

############ Visualization packages
networkx
osmnx==1.1.2
folium
geopandas

############ Annotation packages
nuscenes-devkit==1.1.9

##### General packages
PyYAML
matplotlib
numpy==1.21
pandas
scipy
glob2
progressbar
tqdm
multiprocess


# testing code package
pytest


